requirements.txt
flask
flask-cors
flask-smorest
pyjwt
bcrypt
cryptography